+++ fixtures/attribute-behavior/.gitignore
@@ -10,6 +10,7 @@
 /build
 /public/react.development.js
 /public/react-dom.development.js
+/public/react-dom-server.browser.development.js
 
 # misc
 .DS_Store
