+++ fixtures/attribute-behavior/package.json
@@ -3,6 +3,7 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "file-saver": "^1.3.3",
     "glamor": "^2.20.40",
     "react": "^15.6.1",
     "react-dom": "^15.6.1",
@@ -10,8 +11,7 @@
     "react-virtualized": "^9.9.0"
   },
   "scripts": {
-    "prestart":
-      "cp ../../build/dist/{react,react-dom,react-dom-server.browser}.development.js public/",
+    "prestart": "cp ../../build/dist/{react,react-dom,react-dom-server.browser}.development.js public/",
     "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test --env=jsdom",
