--- fixtures/attribute-behavior/src/react-loader.js
@@ -1,15 +1,0 @@
-export function loadScript(src) {
-  let firstScript = document.getElementsByTagName('script')[0];
-  let scriptNode;
-
-  return new Promise((resolve, reject) => {
-    scriptNode = document.createElement('script');
-    scriptNode.async = 1;
-    scriptNode.src = src;
-
-    scriptNode.onload = () => resolve();
-    scriptNode.onerror = () => reject(new Error(`failed to load: ${src}`));
-
-    firstScript.parentNode.insertBefore(scriptNode, firstScript);
-  });
-}
